fix: cast duration to long long for portable printf format
Fix -Wformat warning on Linux where chrono::milliseconds::rep is long
rather than long long. Cast to long long to match %lld format specifier
portably across platforms.

+++ tests/barrier_stress_test.cpp
@@ -69,7 +69,7 @@ int main(int argc, char* argv[])
     auto end = std::chrono::steady_clock::now();
     auto duration = std::chrono::duration_cast<std::chrono::milliseconds>(end - start);
 
-    std::printf("Barrier stress test completed in %lld ms\n", duration.count());
+    std::printf("Barrier stress test completed in %lld ms\n", static_cast<long long>(duration.count()));
     std::printf("Worker failures: %d\n", worker_failures.load());
     std::printf("Coordinator failures: %d\n", coordinator_failures.load());
